calculate app added
A simple calculator to perform basic arithmetic operations is a good project idea. The calculator consists of an interface to key in the numbers, erase the numbers or delete the whole result. In this calculator, we can include operations such as addition, subtraction, multiplication, and division.

Technology stack

HTML
CSS
JavaScript
Python- Django